+++ Sources/StreamChatSwiftUI/ChatChannel/Composer/MessageComposerView.swift
@@ -217,7 +217,7 @@ public struct MessageComposerView<Factory: ViewFactory>: View, KeyboardReadable 
 }
 
 /// View for the composer's input (text and media).
-public struct ComposerInputView<Factory: ViewFactory>: View {
+public struct ComposerInputView<Factory: ViewFactory>: View, KeyboardReadable {
 
     @EnvironmentObject var viewModel: MessageComposerViewModel
     
@@ -240,6 +240,7 @@ public struct ComposerInputView<Factory: ViewFactory>: View {
     var removeAttachmentWithId: (String) -> Void
 
     @State var textHeight: CGFloat = TextSizeConstants.minimumHeight
+    @State var keyboardShown = false
 
     public init(
         factory: Factory,
@@ -387,11 +388,14 @@ public struct ComposerInputView<Factory: ViewFactory>: View {
         .background(composerInputBackground)
         .overlay(
             RoundedRectangle(cornerRadius: TextSizeConstants.cornerRadius)
-                .stroke(Color(colors.innerBorder))
+                .stroke(Color(keyboardShown ? colors.composerInputHighlightedBorder : colors.innerBorder))
         )
         .clipShape(
             RoundedRectangle(cornerRadius: TextSizeConstants.cornerRadius)
         )
+        .onReceive(keyboardWillChangePublisher) { visible in
+            keyboardShown = visible
+        }
         .accessibilityIdentifier("ComposerInputView")
     }