+++ CHANGELOG.md
@@ -3,6 +3,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### ✅ Added
+- Configuration for highlighted composer border color
+
 ### 🐞 Fixed
 - Improved loading of gallery images